+++ packages/url_launcher/url_launcher_windows/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
+* Fixes an issue where the URL logged would not be unescaped on failure.
 
 ## 3.1.3